SortedArray searches use template argument
This means you can find search with any element that can compare with
the type inside the array. The purpose of this change is to avoid
implicit and useless casts. It makes most sense when thinking of a const
char* being transformed into a String for the sole purpose of being
compared, this can now be avoided.