BufferStream follows feof norm
feof only returns true *after* a read request cannot be completed because the file is ended. BufferStream would signal end of file even if the last read request was successful but resulted in the end of the file being reached. To unify the behaviours it is simpler to make BufferStream act like a C stream, however I do not like that way of thinking about the end of a stream. I'll need to find a nice way to make C streams behave as I want.